# TS compilation perf: steal faster extendShape type from tRPC