@colinhacks 2023-10-04
I feel like there are edge cases where dropping `flatten` breaks the inferred type signature...but I can't find them.

I'll do a little more testing then merge. Thanks!