@jussisaurio 2023-10-10
FWIW I can try to run some perf tests on how much keeping `flatten` has an impact. I'm guessing though that replacing `identity` with something like `& {}` in the definition of `flatten` would also make it perform better due to less instantiations