@jussisaurio 2023-10-10
Used this: https://github.com/jussisaurio/zod-ts-perftest (there are already pregenerated zod schemas in the repo)

Running `npm run build-bench`, three runs each:

master: 4.51s, 4.56s, 4.51s
PR #2839 (this one): 3.69s, 3.61s, 3.72s
PR #2839 + flatten added back: 3.77s, 3.77s, 3.72s
PR #2845: 2.29s, 2.20s, 2.21s
Both PRs: 2.14s, 2.14s, 2.16s

Not too much of a difference at all between keeping or removing `flatten`, but for some reason this benchmark is really sensitive to the changes in #2845 😲 I'm sure this particular benchmark is not very well rounded (`zod-ts-perftest` only generates objects, strings, numbers, booleans, extends and omits).