@psychedelicious 2024-02-12
We use extend quite a bit. This PR is a marked improvement.

zod 3.22.4:
```
npx tsc --noEmit --extendedDiagnostics --incremental false

Files:                         3394
Lines of Library:             40241
Lines of Definitions:        310600
Lines of TypeScript:          73937
Lines of JavaScript:              0
Lines of JSON:                    0
Lines of Other:                   0
Identifiers:                 592152
Symbols:                     950503
Types:                       311682
Instantiations:             8492589
Memory used:               1151591K
Assignability cache size:    273407
Identity cache size:          23254
Subtype cache size:            5426
Strict subtype cache size:     7230
I/O Read time:                0.04s
Parse time:                   0.65s
ResolveModule time:           0.20s
ResolveTypeReference time:    0.00s
ResolveLibrary time:          0.01s
Program time:                 1.02s
Bind time:                    0.34s
Check time:                   7.75s
printTime time:               0.00s
Emit time:                    0.00s
Total time:                   9.12s
```

zod 3.22.4 with this patch:
```
npx tsc --noEmit --extendedDiagnostics --incremental false

Files:                         3394
Lines of Library:             40241
Lines of Definitions:        310610
Lines of TypeScript:          73937
Lines of JavaScript:              0
Lines of JSON:                    0
Lines of Other:                   0
Identifiers:                 592160
Symbols:                     950391
Types:                       310870
Instantiations:             6238717
Memory used:               1058367K
Assignability cache size:    273343
Identity cache size:          23254
Subtype cache size:            5426
Strict subtype cache size:     7230
I/O Read time:                0.04s
Parse time:                   0.66s
ResolveModule time:           0.21s
ResolveTypeReference time:    0.00s
ResolveLibrary time:          0.01s
Program time:                 1.03s
Bind time:                    0.35s
Check time:                   7.32s
printTime time:               0.00s
Emit time:                    0.00s
Total time:                   8.70s
```

The time to check the whole app isn't a big deal, but VSCode's TS intellisense is a pain. It slowed down substantially after I rewrote a bunch of schemas using extend. We have a number of  nested extend schemas in big unions.