@colinhacks 2024-04-20
Merged a variant in https://github.com/colinhacks/zod/pull/2845

Thanks!